Extract config loader into module
