+++ packages/config/.gitignore
@@ -1,0 +1,5 @@
+/dist
+/lib
+/standalone
+
+stats.html
