+++ packages/config/package.json
@@ -1,0 +1,35 @@
+{
+  "name": "@zudoku/config",
+  "type": "module",
+  "version": "0.14.1",
+  "keywords": [
+    "zudoku"
+  ],
+  "description": "Zudoku configuration loader",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/zuplo/zudoku",
+    "directory": "packages/config"
+  },
+  "author": "Zuplo <support@zuplo.com>",
+  "license": "MIT",
+  "exports": {
+    ".": "./dist/index.js"
+  },
+  "files": [
+    "dist"
+  ],
+  "scripts": {
+    "build": "tsc"
+  },
+  "engines": {
+    "node": ">=20"
+  },
+  "dependencies": {
+    "tsx": "4.19.1"
+  },
+  "devDependencies": {
+    "@types/node": "20.16.11",
+    "typescript": "5.6.3"
+  }
+}
