+++ packages/config/src/index.ts
@@ -1,0 +1,5 @@
+export {
+  ConfigWithMeta,
+  loadZudokuConfig,
+  zudokuConfigFiles,
+} from "./config.js";
