+++ packages/zudoku/package.json
@@ -173,6 +173,7 @@
     "@types/react": "18.3.11",
     "@types/react-dom": "18.3.1",
     "@vitejs/plugin-react": "4.3.1",
+    "@zudoku/config": "workspace:*",
     "@zudoku/httpsnippet": "10.0.9",
     "@zudoku/react-helmet-async": "2.0.4",
     "autoprefixer": "10.4.20",
@@ -219,7 +220,6 @@
     "strip-ansi": "7.1.0",
     "tailwind-merge": "2.5.4",
     "tailwindcss": "3.4.13",
-    "tsx": "4.19.1",
     "ulidx": "2.4.1",
     "unist-util-visit": "5.0.0",
     "urql": "4.1.0",
