+++ packages/zudoku/src/vite/plugin-config.ts
@@ -1,14 +1,13 @@
 import { type Plugin } from "vite";
-import { getConfigFilePath } from "./config.js";
 
-const viteConfigPlugin = ({ rootDir }: { rootDir: string }): Plugin => {
+const viteConfigPlugin = (configPath: string): Plugin => {
   const virtualModuleId = "virtual:zudoku-config";
 
   return {
     name: "zudoku-config-plugin",
     resolveId(id) {
       if (id === virtualModuleId) {
-        return getConfigFilePath(rootDir);
+        return configPath;
       }
     },
   };
